+++ NEWS.md
@@ -16,6 +16,7 @@
 - RStudio now supports `LuaLaTeX` to compile Sweave/Rnw documents. (#13812)
 - Better error message when user preferences fail to save due to folder permissions. (#12974)
 - Update Electon Forge to 6.4.2 and Webpack to 5.89.0. (rstudio-pro#5383)
+- RStudio now supports pasting of file paths for files copied to the clipboard. (#4572)
 
 #### Posit Workbench
 - Removed link for opening sessions in RStudio Desktop Pro from Session Info dialog. (rstudio-pro#5263)